This is my initial commit of my node-irrigation.js project. An irrigation controller for Linux and the Elexol EtherIO24 board.
My initial work was done under git control with a Bitbucket project. A
project called t (https://bitbucket.org/linuxha/t). Just to get a better
hands on feel for git version control. During that time a great many
undocumented features were added and much of the software architecture
decided apon. A whle lot of work is needed to get this out of the intial
beta release and into the stable release.