Added bin/git_status.sh and tests/tests.js
I've forgotten why I added git_status.sh (must have been something I was learning about with git) but the tests.js was so I could one day run tests with mocha to test the readiness of this program. At this time I haven't got a clue how to perform and level of tests on this.

[ticket: X]